Include a reference to the Ampersand Documentation
This tools documentation is the default when you go to `https://github.com/AmpersandTarski/`. People that end up here but just want to start using ampersand may well be interested in the documentation of Ampersand itself and how to install the stuff for creating protoyptes etc. I added a paragraph for users of this kind.